+++ apps/delivery-options/src/utils/hasPickupForCarrier.ts
@@ -1,0 +1,9 @@
+import {type CarrierCapability, type CarrierIdentifier, CarrierSetting} from '@myparcel-dev/do-shared';
+import {getResolvedValue} from './getResolvedValue';
+
+export const hasPickupForCarrier = (cap: CarrierCapability, carrierIdentifier?: CarrierIdentifier): boolean => {
+  return (
+    cap.deliveryTypes.includes('PICKUP_DELIVERY') &&
+    Boolean(getResolvedValue(CarrierSetting.AllowPickupLocations, carrierIdentifier))
+  );
+};
