+++ apps/delivery-options/src/utils/index.ts
@@ -6,6 +6,7 @@ export * from './createGetDeliveryOptionsParameters';
 export * from './createLatLngParameters';
 export * from './createNextDate';
 export * from './createUtcDate';
+export * from './getCapabilityDeliveryTypes';
 export * from './getConfigFromWindow';
 export * from './getConfigPriceKey';
 export * from './getDefaultAddress';
@@ -13,6 +14,8 @@ export * from './getDeliveryTypePrice';
 export * from './getResolvedCarrier';
 export * from './getResolvedDeliveryType';
 export * from './getResolvedValue';
+export * from './hasDeliveryForCarrier';
+export * from './hasPickupForCarrier';
 export * from './hasSlotContent';
 export * from './language';
 export * from './padTime';
